requirements.txt
flask~=1.1.2
Werkzeug~=1.0.1
Pillow~=6.2.1